+++ include/logger.h
@@ -1,0 +1,20 @@
+/******************************************************************************
+ *
+ * @file        logger.h
+ *
+ * @author      Icko X. Iben
+ *
+ * @date        Friday, 31 May 2024
+ *
+ * @brief       Defines our log file
+ *
+ *****************************************************************************/
+
+#ifndef LOGGER_H
+#define LOGGER_H
+
+#include <stdio.h>
+
+extern FILE *logFile;
+
+#endif