+++ docs/changelog.md
@@ -5,3 +5,20 @@ Changelog for the wtfsh
 05/31/2024
 
 * Project created
+
+---
+
+06/08/2024
+
+* `nsh_getline` and `nsh_tokenize` working
+
+---
+
+06/09/2024
+
+* Builtin function pointers added
+* DEBUG prints added
+* Logger.h initial
+    * TODO: add DEBUG prints to log file
+* `cd` implementation finished
+* `help` implementation finished